--- coprocess/python/proto
@@ -1,1 +1,0 @@
-../bindings/python/
