--- coprocess_python_api.c
@@ -1,1 +1,0 @@
-coprocess/python/tyk/gateway.c
