+++ load_balancer/round_robin.go
@@ -1,0 +1,16 @@
+package load_balancer
+
+import "sync/atomic"
+
+type RoundRobin struct {
+	pos uint32
+}
+
+func (r *RoundRobin) WithLen(len int) int {
+	if len < 1 {
+		return 0
+	}
+	// -1 to start at 0, not 1
+	cur := atomic.AddUint32(&r.pos, 1) - 1
+	return int(cur) % len
+}
