+++ rpc_backup_handlers.go
@@ -89,7 +89,7 @@ func doLoadWithBackup(specs []*APISpec) {
 	}
 
 	// Reset the JSVM
-	GlobalEventsJSVM.Init()
+	GlobalEventsJSVM.Init(&globalConf, &APIHandle)
 	log.Warning("[RPC Backup] --> Initialised JSVM")
 
 	newRouter := mux.NewRouter()
