--- sds.c
@@ -1,1 +1,0 @@
-coprocess/sds/sds.c
