+++ session/session_state.go
@@ -1,4 +1,4 @@
-package auth
+package session
 
 import (
 	"fmt"
