+++ session_state.go
@@ -1,18 +1,18 @@
 package main
 
 import (
-	"github.com/TykTechnologies/tyk/auth"
+	"github.com/TykTechnologies/tyk/session"
 )
 
 // This stub is part of a refactor, please see auth/session_state for the Session objects aliased below.
 
 const (
-	HashPlainText = auth.HashPlainText
-	HashBCrypt    = auth.HashBCrypt
+	HashPlainText = session.HashPlainText
+	HashBCrypt    = session.HashBCrypt
 )
 
-type AccessSpec = auth.AccessSpec
-type AccessDefinition = auth.AccessDefinition
-type SessionState = auth.SessionState
+type AccessSpec = session.AccessSpec
+type AccessDefinition = session.AccessDefinition
+type SessionState = session.SessionState
